+++ pkg/defaultinvariants/types.go
@@ -3,7 +3,7 @@ package defaultinvariants
 import (
 	"fmt"
 
-	"github.com/openshift/origin/pkg/invariants/disruptionpodnetwork"
+	"github.com/openshift/origin/pkg/invariants/network/disruptionpodnetwork"
 
 	"github.com/openshift/origin/pkg/invariants"
 	"github.com/openshift/origin/pkg/invariants/authentication/legacyauthenticationinvariants"
@@ -65,14 +65,13 @@ func NewInvariantsFor(clusterStabilityDuringTest ClusterStabilityDuringTest) inv
 func newDefaultInvariants() invariants.InvariantRegistry {
 	invariantTests := invariants.NewInvariantRegistry()
 
-	invariantTests.AddInvariantOrDie(disruptionpodnetwork.InvariantName, disruptionpodnetwork.JIRAOwner, disruptionpodnetwork.NewPodNetworkAvalibilityInvariant())
-
 	invariantTests.AddRegistryOrDie(newUniversalInvariants())
 
 	invariantTests.AddInvariantOrDie("image-registry-availability", "Image Registry", disruptionimageregistry.NewAvailabilityInvariant())
 
 	invariantTests.AddInvariantOrDie("apiserver-availability", "kube-apiserver", disruptionlegacyapiservers.NewAvailabilityInvariant())
 
+	invariantTests.AddInvariantOrDie(disruptionpodnetwork.InvariantName, "Network / ovn-kubernetes", disruptionpodnetwork.NewPodNetworkAvalibilityInvariant())
 	invariantTests.AddInvariantOrDie("service-type-load-balancer-availability", "NetworkEdge", disruptionserviceloadbalancer.NewAvailabilityInvariant())
 	invariantTests.AddInvariantOrDie("ingress-availability", "NetworkEdge", disruptioningress.NewAvailabilityInvariant())
