+++ pkg/invariants/network/disruptionpodnetwork/namespace.yaml
