+++ pkg/invariants/network/disruptionpodnetwork/pod-network-poller-deployment.yaml
